Reuse babel between prepare.js and watch-js
