+++ scripts/watch-js.js
@@ -1,0 +1,3 @@
+const { babelify } = require('./compile-js');
+
+babelify({ watch: true, silent: false });
